requirements.txt
pandas>=1.3.0
requests>=2.26.0
beautifulsoup4>=4.9.3
selenium>=4.10.0
webdriver-manager>=4.0.0
cfbd>=0.1.0
python-dotenv>=1.0.0